+++ src/Response/Model/SectionMedia.php
@@ -7,14 +7,19 @@ use InstagramAPI\AutoPropertyMapper;
 /**
  * SectionMedia.
  *
+ * @method Clip getClips()
  * @method Item getMedia()
+ * @method bool isClips()
  * @method bool isMedia()
+ * @method $this setClips(Clip $value)
  * @method $this setMedia(Item $value)
+ * @method $this unsetClips()
  * @method $this unsetMedia()
  */
 class SectionMedia extends AutoPropertyMapper
 {
     const JSON_PROPERTY_MAP = [
         'media'  => 'Item',
+        'clips'  => 'Clip',
     ];
 }